+++ aidial_adapter_openai/embeddings/vllm/qwen3_vl_api.py
@@ -8,14 +8,43 @@ import asyncio
 from dataclasses import dataclass
 
 from aidial_sdk.embeddings.request import EmbeddingsRequest
-from aidial_sdk.embeddings.response import EmbeddingResponse
+from aidial_sdk.embeddings.response import Embedding, EmbeddingResponse, Usage
 from pydantic import BaseModel
 
 from aidial_adapter_openai.embeddings.vllm.client import post_upstream
 from aidial_adapter_openai.embeddings.vllm.media import image_content_part
-from aidial_adapter_openai.embeddings.vllm.openai_api import merge_fanout
 from aidial_adapter_openai.utils.auth import OpenAICreds
 from aidial_adapter_openai.utils.resource.base import Resource
+
+
+def merge_fanout(model: str, responses: list[dict]) -> EmbeddingResponse:
+    vectors: list[Embedding] = []
+    prompt_tokens = 0
+    total_tokens = 0
+    for idx, raw in enumerate(responses):
+        parsed = EmbeddingResponse.model_validate(raw)
+        item = (
+            parsed.data[0]
+            if parsed.data
+            else Embedding(embedding=[], index=idx)
+        )
+        vectors.append(Embedding(embedding=item.embedding, index=idx))
+        if parsed.usage:
+            prompt_tokens += parsed.usage.prompt_tokens
+            total_tokens += parsed.usage.total_tokens
+
+    total_tokens = total_tokens or len(vectors)
+    prompt_tokens = prompt_tokens or total_tokens
+
+    return EmbeddingResponse(
+        model=model,
+        data=vectors,
+        usage=Usage(
+            prompt_tokens=prompt_tokens,
+            total_tokens=total_tokens,
+        ),
+    )
+
 
 _QWEN3_VL_DEFAULT_INSTRUCTION = "Represent the user's input."
